Integration with Kratos API
This commit introduces a Kratos API wrapper that implements the user management functionalities required by the task management API. For now there is just one feature needed - to get a corresponding user identity given an input authentication token. With this functionality each new task that is created will have a reference to the ID of the user that created it.